requirements.txt
dragonfly2>=0.34.0
pillow>=6.2.2
tomlkit>=0.5.11
future>=0.18.2
mock>=3.0.5
appdirs>=1.4.3
scandir>=1.10.0
pyvda>=0.2.3
PySide2>=5.14
six
